Fixed decimal number format.
Transform to string decimal and double arrays using
CultureInfo.InvariantCulture in order to not use comma in some
locales when preparing the javascript code.

This fixes cases where the output javascript is generated from
a Series of data made of a double[] on locales with comma
as decimal separator the generated javascript code uses
comma as decimal separator, which is wrong and will yield
to wrong results: eg: [0,2394 , 0,24932] instead of [0.2394 , 0.24932].
Differently with locales like USA this works correctly.